818 documentation cleanup (#869)
Closes #818 

With this PR we aimed to explain `teal.code` usage in `Adding support
for Reporting to custom modules` vignette. This vignette demonstrates
the usage of `teal.reporter` modules with `teal` apps. The example with
`teal.code` is not necessary for the vignette however it shows an
advanced combination of usage of `teal.code` and `teal.reporter` hence I
decided not to remove `teal.code` snippets from this article.

There are also `teal.code` snippets in another vignette (`Creating
Custom Modules`), so removal from this one would not remove the Suggests
dependency for the `teal.code` package anyway.

I also reviewed the text and tried to polish it. 

Lastly, there examples in `R/module_tabs_with_filters.R` were updated
during https://github.com/insightsengineering/teal/issues/870

---------

Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: Aleksander Chlebowski <114988527+chlebowa@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>