update install guide; add getting help; pkg name lowercase (#890)
- update install guide - added back install from GH as this is still a
standard way of installing non-CRAN packages. I added install via `pak`
and not `remotes` because this is also standard for pkgs from tidyverse
or r-lib.
- change urls to teal gallery and tlg-catalog
- added Getting Help section with reference to slack channel
- replace `browseVignettes(package = "teal")` with url as I believe that
most of users will read README from within pkgdown webpage.
- replace "Teal" with "teal" to keep consistent package name
- UPDATE: added "Usage section" with gif recording and move all "Please
refer to ... for more" under that section

Please add your feedback. I would like to treat this PR as a reference
to the issue for other package updates.

---------

Signed-off-by: Pawel Rucki <12943682+pawelru@users.noreply.github.com>
Co-authored-by: Aleksander Chlebowski <114988527+chlebowa@users.noreply.github.com>
Co-authored-by: kartikeya kirar <kirar.kartikeya1@gmail.com>