Fix the integration tests by alternative  (#1237)
Changes:

1. The container in the integration test does not seem to evaluate the
JS code to check visibility when `app_driver$is_visible` is called with
`visibility_property = TRUE`
2. Increasing the timeout threshold for `wait_for_page_stability`. For
an unknown reason this timeout does not work in certain containers.
Initial timeout (500ms) was set as a 2x of what was needed in my local
machine. But looks like we need at least 800 ms to work in some loaded
conditions. So, setting it to 2000 ms now.