Improve progress bars (#1236)
Related to #1234.

Update progress bars to improve user experience.

- `modules_datasets` receives optional `progress` argument that accepts
a `Progress` object
- `ui_nested_tabs` and `ui_tabs_with_filters` receive `progress`
argument that accepts a `Progress` object
- rewrote progress bar management in `srv_teal`

Added a large app for testing in `__stress_test.R`, to be removed before
merging.

#### NOTE

Once the module UI is prepared, there is still a delay before the app
starts, so the progress bar can be improved further. So far I was unable
to identify the reason for the delay.