fix `noSuggests` check (#1259)
This is my testing branch of r-hub workflows. If you see this as
non-draft that means that I have removed my workflow file and therefore
making it ready to be merged (this is do be done separately on the next
propagation round).

The result of my tests are here:
https://github.com/insightsengineering/teal/actions/workflows/rhub.yaml

One of the checks is "noSuggests" check and it runs the check without
suggested packages. This fixes encountered errors.

Please check if the downstream packages requires similar updates. If so,
please kindly do this or let me know.