Rename  teal_data_module.R functions (#1430)
# Pull Request


Renamed parameters to data, and renamed data_rv, filtered_teal_data and teal_data parameter to data. This simplifies the different parameters used across teal and instead of having 3 diferents parameters we will have one. This will avoid affecting the end users and issues with documentation diverging on different places.

Internal module objects were also renamed to be more consistent. 

The data parameter may refer to different classes, but developers should rely on the assertions to know which class is accepted instead of the name.
Fixes #1323 

Due to the renaming I had to change some tests. 

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Marcin <133694481+m7pr@users.noreply.github.com>