link modules by id (#1555)
Modules tab-content and buttons in a dropdown are linked by the id of
the `modules` object.
No need to parse the groups recursively and keep informations from each
ancestor, we just need to create unique identifier for given module
which will be a link between independently called modules.

todo:
- two identical modules will create duplicated hash
- initial `active_module_id` needs to be on the first (recursive)
module.

---------

Co-authored-by: vedhav <vedhaviyash4@gmail.com>