Update gif instead of shinylive (#1569)
# Pull Request

<!--- Replace `#nnn` with your issue link for reference. -->
Replacement for https://github.com/insightsengineering/teal/pull/1490


Tried to create a `.gif` but it either was too small on the screen or it
weighted much more than the video.
The raw `.mp4` is not displayed on the github branch so I went with the
`.webp` that was previously used.