Update vignettes with the latest screenshots (#1574)
Changes:

1. Updated the gif recording for the main readme and pkgdown site
homepage as it contains the older `sessionInfo` link at the bottom +
improve the gif quality.
2. Updated the screenshots from the bootstrap vignette as they too
expose the older `sessionInfo` link + they show local changes like
showing the lockfile download button that is only visible on setting the
option which will not be the case for the majority of the users copy
pasting the example on their local.
3. Updated the show R code and reporter screenshots with meaningful
arrows indicating the relationships between different screens, it was
not clear before.
4. Moved all the image assets inside `vignettes` directory for easy
reference and maintenance.