Update report in custom modules vignette for new teal.reporter (#1593)
# Pull Request

Fixes #1591

Update the vignette to explain how teal.reporter can be used to generate
the report on modules.


There might be room for improvement on the vignette as this is my first
module using this new features.
Specially this section of code doesn't seem self explanatory:

```r
 teal.reporter::teal_card(obj) <-
            c(
              teal.reporter::teal_card("# Module with reporting"),
              teal.reporter::teal_card(obj),
              teal.reporter::teal_card("## Module's code")
            )
```

It is an object of class `teal_report` (of a single module!) being
modified by the function `teal_card` to add `teal_card`s. But I might
have got this wrong.

Another thing I'm not confident is on having to use
`module()[[input$dataname]]` for the `renderPrint`. After having to add
a line to print the table selected for the report.


I removed lot's of apps that showed intermediate work, as I think now it
can be done in two steps:
 1. Adding a card
 2. Returning it to be used by teal

Awesome work that went into this! 
In general, it felt very easy to set it up once I got around the
`teal_reporter` and `teal_card` classes seeing other modules on TMG. But
let me know if I missed a clearer description.

---------

Signed-off-by: Lluís Revilla <185338939+llrs-roche@users.noreply.github.com>
Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>
Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Marcin <133694481+m7pr@users.noreply.github.com>