Copilot Coding Agent Setup for Github Actions environment (#1659)
When GitHub Copilot Coding Agent is run, it is run in GitHub Actions
environment.
This yaml file, specify the setup that will be used during the execution
of the job.

Coding Agent will work in the same docker conatiner as is used for
CI/CD.
It also verifies that R is installed and presents R session, before it
starts any job.

This way, we allow coding agent to use R and we use the same environment
as in our testing scenarios during CI/CD

Fixes
- https://github.com/insightsengineering/coredev-tasks/issues/678

Partner to 
- https://github.com/insightsengineering/teal/pull/1628

Signed-off-by: Marcin <133694481+m7pr@users.noreply.github.com>